Revert "wait for serial connection"
This reverts commit 0c7874e8706f7c01e4968082678508f15212611a.

Many users base their own code off of examples. If they are trying to
make a robot, this will lead to trouble. Because the program will wait
for the serial connection to open, the robot will not move. For this
reason `while (!Serial) ;` is not a good solution in this case.