update "Wait until user presses GO..."
A majority of the EVShield tests and examples use a statement to "Wait
until user presses GO button to continue the program."  With this new
approach the message will be repeated about once a second. This is
preferable, for if the user had not had the serial terminal open they
would not have seen the message. This is more significant on the
Arduino/Genuino 101 where the sketch does not restart when the serial
terminal is open.
A couple of approaches [were
considered](https://gist.github.com/sethten/4c6a0fc9618b81e8b30a6b9c9a2d8135).
millis() returns the number of milliseconds since the sketch started. We
test the remainder of this value when divided by one thousand. Testing
for equality with zero would not be as successful as this boolean
expression is only evaluated when the loop reaches that point, which may
not be exactly on a thousandth millisecond. It seems checking for the
range of values zero, one, or two results in the message being printed
about once per second. Also because we <3 our users.