update "Wait until user presses GO..." (#20)
* wait for serial connection

Instead of waiting a static number of milliseconds, or not waiting at
all, wait for the serial connection to be open before continuing and
trying to send serial communication. The arduino.cc
[guide](https://www.arduino.cc/en/Guide/Arduino101#toc6) for the Arduino
101 suggests this method. using `while (!Serial) ;`

* Revert "wait for serial connection"

This reverts commit 0c7874e8706f7c01e4968082678508f15212611a.

Many users base their own code off of examples. If they are trying to
make a robot, this will lead to trouble. Because the program will wait
for the serial connection to open, the robot will not move. For this
reason `while (!Serial) ;` is not a good solution in this case.

* update "Wait until user presses GO..."

A majority of the EVShield tests and examples use a statement to "Wait
until user presses GO button to continue the program."  With this new
approach the message will be repeated about once a second. This is
preferable, for if the user had not had the serial terminal open they
would not have seen the message. This is more significant on the
Arduino/Genuino 101 where the sketch does not restart when the serial
terminal is open.
A couple of approaches [were
considered](https://gist.github.com/sethten/4c6a0fc9618b81e8b30a6b9c9a2d8135).
millis() returns the number of milliseconds since the sketch started. We
test the remainder of this value when divided by one thousand. Testing
for equality with zero would not be as successful as this boolean
expression is only evaluated when the loop reaches that point, which may
not be exactly on a thousandth millisecond. It seems checking for the
range of values zero, one, or two results in the message being printed
about once per second. Also because we <3 our users.