Change if-else chain to a switch
An if-else chain that does an `==` check on a number is the exact same as a switch in c++, there's no reason to use if.
I also took the liberty of wrapping some of the server functions in a macro for easier typing on my part. The macros expand on compile-time anyway, so it shouldn't make much difference performance-wise.